refactor(core): transformer option name
